+++ sdk/src/main/java/com/walmartlabs/concord/sdk/Constants.java
@@ -220,6 +220,11 @@ public class Constants {
          * Use the trigger's commit id to start the process.
          */
         public static final String USE_EVENT_COMMIT_ID = "useEventCommitId";
+
+        /**
+         * Exclusive group.
+         */
+        public static final String EXCLUSIVE_GROUP = "exclusive";
     }
 
     /**